Avoid crash caused by long messages without spaces
If a long message contains no spaces (like a link) lastIndexOf will return -1. In that case the call to substring() will throw a IndexOutOfBoundsException and crash the game.

To fix, we check whether the lastIndexOf found any spaces. If it did use the old logic. If it doesn't then just append ... to the end of the string. Note that this means that we can end up with something like mylongmessage... without a space between the message and the ellipsis.